--- ganghee/MyStudy2/app/src/main/java/com/example/mystudy/network/AppController.kt
@@ -1,20 +1,0 @@
-package com.example.mystudy.network
-
-import retrofit2.Retrofit
-import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
-import retrofit2.converter.gson.GsonConverterFactory
-
-object RetrofitClient {
-    private var ourInstance : Retrofit?= null
-    val instance : Retrofit
-        get() {
-            if (ourInstance == null) {
-                ourInstance = Retrofit.Builder()
-                    .baseUrl("https://api.upbit.com/v1/")
-                    .addConverterFactory(GsonConverterFactory.create())
-                    .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
-                    .build()
-            }
-            return ourInstance!!
-        }
-}
