+++ ganghee/MyStudy2/app/src/main/java/com/example/mystudy/network/UpbitApi.kt
@@ -1,7 +1,5 @@
 package com.example.mystudy.network
 
-import com.example.mystudy.network.MarketResponse
-import com.example.mystudy.network.TickerResponse
 import io.reactivex.Single
 import retrofit2.http.GET
 import retrofit2.http.Query
@@ -16,6 +14,4 @@ interface UpbitApi {
     fun getTickers(
         @Query("markets") markets: String
     ): Single<List<TickerResponse>>
-
-
 }
